% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codo}
\alias{codo}
\title{Title
Codon Data
Description
This is file storing the data for the project.}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codo
}
\description{
Title
Codon Data
Description
This is file storing the data for the project.
}
\keyword{datasets}
