% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codoInSeq.R
\name{codoInSeq}
\alias{codoInSeq}
\title{codoInSeq}
\usage{
codoInSeq(sequence, start = 1)
}
\arguments{
\item{sequence}{this is the base sequence to get codons from}

\item{start}{this is point in sequence from which we start getting codons}
}
\value{
these are the codons in sequence from selected starting point
}
\description{
codoInSeq
}
\examples{
x <- "UUUUCUUAUUGUUUCUCCUACUGCUUAUCAUAAUGAUUGUCGUAGUGGCUUCCUCAUCGUCUCCCCCAC"
codoInSeq(x, start = 1)
codoInSeq(x, start = 2)
}
