% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_data}
\alias{codon_data}
\title{Codon to Amino Acid Mapping}
\format{
A named vector with 64 elements:
\describe{
  \item{UUU}{F}
  \item{UCU}{S}
  \item{UAU}{Y}
  ...
}
}
\usage{
codon_data
}
\description{
A named vector containing the mapping from codons to their corresponding one-letter amino acid codes.
}
\keyword{datasets}
