% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Standard codon table}
\format{
A dictionary containing 64 key-value pairs:
\describe{
  \item{key}{3 letter codon}
  \item{value}{one letter code}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table
}
\description{
A list of codons with their corresponding one letter codes.
}
\keyword{datasets}
