% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_to_protein_4.R
\name{codons_to_proteins}
\alias{codons_to_proteins}
\title{Codons to sequence of aa}
\usage{
codons_to_proteins(codons)
}
\arguments{
\item{codons}{character vector of RNA codons}
}
\value{
string representing the translated aa sequence
}
\description{
Translates a vector of codons into an aminoacid sequence
}
\details{
This function takes a vector of RNA codons and translates them into a
single aa sequence, using the codon table
}
\examples{
translate_codons("AGU")

codons <- c("AUG", "UUU", "AGA")
translate_codons(codons)

}
