% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_rna.R
\name{convert_to_rna}
\alias{convert_to_rna}
\title{Convert to DNA to RNA}
\usage{
convert_to_rna(dna)
}
\arguments{
\item{dna}{string containing a DNA strand in the format 'ATGC'}
}
\value{
returns the matching RNA strand after translation
}
\description{
Convert to DNA to RNA
}
\examples{
convert_to_rna('TAGTCGA')
}
