% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Transcription}
\usage{
dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{character string of DNA sequence}
}
\value{
character string of RNA sequence with U instead of T
}
\description{
Conversion of a DNA sequence to a RNA
The function converts a DNA sequence (input) to a RNA sequence by
replacing the T with a U
}
\examples{
dna_to_rna("ATGC")
dna_to_rna(c("ATGCTTAGGCGT", "TTTGGGAAA"))


}
