% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Extraction of codons}
\usage{
extract_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A character string representing the DNA or RNA sequence from which codons are to be extracted.}

\item{start}{An integer indicating the position in the sequence to start extraction (default is 1).}
}
\value{
A character vector containing the extracted codons from the sequence.
}
\description{
This function extracts codons (triplets of characters) from a given DNA or RNA sequence,
starting from a specified position.
}
\examples{
# Example usage
extract_codons("ATGCGATAGCTA", start = 1)
# Expected output: c("ATG", "CGA", "TAG")

extract_codons("ATGCGATAGCTA", start = 2)
# Expected output: c("TGC", "GAT", "AGC")
}
