% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DNA.R
\name{generate_DNA}
\alias{generate_DNA}
\title{Generate a DNA sequence}
\usage{
generate_DNA(DNA_length)
}
\arguments{
\item{DNA_length}{The size of the DNA sequence to be generated (int)}
}
\value{
A DNA sequence of the desired length (str)
}
\description{
Generate a DNA sequence
}
\examples{
DNA_seq_50 <- generate_DNA(DNA_length = 50)
DNA_seq_50

}
