% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nucleotide_Count.R
\name{Nucleotide_Counts}
\alias{Nucleotide_Counts}
\title{Nucleotide Count Plot}
\usage{
Nucleotide_Counts(Nuc_seq)
}
\arguments{
\item{Nuc_seq}{A character string representing a DNA or RNA sequence.}
}
\value{
A bar plot of nucleotide counts of the input sequence.
}
\description{
This function takes a DNA or RNA sequence as input and produces a bar plot
showing the counts of each nucleotide in the sequence.
}
\examples{
# Example DNA sequence
Nuc_seq <- "AGCTTAGG"
Nucleotide_Counts(Nuc_seq)
}
