% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_character_frequency.R
\name{plot_character_frequency}
\alias{plot_character_frequency}
\title{Plot the frequency of each unique character in a given string}
\usage{
plot_character_frequency(input_string)
}
\arguments{
\item{input_string}{The string whose character frequencies are to be plotted.}
}
\value{
A ggplot2 object representing the plot.
}
\description{
This function takes a string as input and plots a bar chart showing the frequency
of each unique character in the string. The plot is generated using ggplot2.
}
\examples{
plot_character_frequency("hello")
}
