% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_aa.R
\name{plot_count_aa}
\alias{plot_count_aa}
\title{Plot number of aa in a sequence}
\usage{
plot_count_aa(sequence_aa)
}
\arguments{
\item{sequence_aa}{string of amino acids}
}
\value{
plot of number of aminoacids in a sequence
}
\description{
la la la
}
