% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seq_dna.R
\name{random_seq_dna}
\alias{random_seq_dna}
\title{Random DNA sequence generator with size input}
\usage{
random_seq_dna(dna_size)
}
\arguments{
\item{dna_size}{Size of the DNA sequence that will be generated}
}
\value{
It is a single character string formed by joining all elements of the random generated DNA with size dna_size
}
\description{
Random DNA sequence generator with size input
}
\examples{
random_seq_dna(dna_size = 10)
random_seq_dna(dna_size = 50)
}
