% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription from DNA to mRNA}
\usage{
transcription(DNA_seq)
}
\arguments{
\item{DNA_seq}{The DNA sequence to transcribe}
}
\value{
The transcribed mRNA sequence
}
\description{
Transcription from DNA to mRNA
}
\examples{
DNA_seq <- 'AAACCTGGACTGAACCTTTGGAACATG'
mRNA_seq <- transcription(DNA_seq)
mRNA_seq

}
