% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codons.R
\name{translate_codons}
\alias{translate_codons}
\title{Translate RNA Codons to Protein Sequence}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{A character vector of RNA codons (e.g., c("AUG", "UUU", "GGC")).}
}
\value{
A string representing the translated protein sequence.
}
\description{
This function translates a vector of RNA codons into a corresponding amino acid sequence.
}
\examples{
translate_codons(c("AUG", "UUU", "GGC"))
# Output: "MFG"

translate_codons(c("UUU", "UCU", "UAU"))
# Output: "FSY"
}
