% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation from codons to Amino Acid sequence}
\usage{
translation(codons)
}
\arguments{
\item{codons}{A list of codons to translate}
}
\value{
The AA sequence (str)
}
\description{
Translation from codons to Amino Acid sequence
}
\examples{

codons <- c("AAA", "CCU", "GGA", "CUG", "AAC", "CUU", "UGG", "AAC", "AUG")
AA_seq <- translation(codons)
AA_seq

}
