% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{DNA sequence to RNA}
\usage{
dna_to_rna(sequence)
}
\arguments{
\item{sequence}{DNA sequence}
}
\value{
Input DNA sequence converted into RNA
}
\description{
DNA sequence to RNA
}
\examples{
Example usage:
dna_to_rna('AATTGAC')
Output:
'AAUUGAC'

}
